Moved ParseStringConverter to Json folder.
