+++ Tricycle.Media.FFmpeg/Models/FFprobe/Stream.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 using Newtonsoft.Json;
-using Tricycle.Media.FFmpeg.Serialization;
+using Tricycle.Media.FFmpeg.Serialization.Json;
 
 namespace Tricycle.Media.FFmpeg.Models.FFprobe
 {
